Bug 23434: Hold confirmation dialog problem if HoldsAutoFill is enabled
This patch corrects a problem with hold confirmation dialogs on the
checkin page when the HoldsAutoFill system preference is enabled and the
item being checked in has a hold at another library.

The confirmation dialog is converted to a modal to match other
confirmation information.

As part of this process two whole sections of the template have been
removed because they were redundant.

To test, apply the patch and test the following circumstances both with
HoldsAutoFill enabled AND disabled:

 - A hold for a patron at your library
 - A hold for a patron not at your library
   - With HoldsAutoFill ON, both the "OK" and "Print slip and confirm"
     buttons should dismiss the modal without reloading the page.
 - A hold which is already marked 'Waiting'
 - An item which needs to be transferred
 - An item which has already been transferred

In all cases the confirmation modal should appear with correct
information.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>