Bug 22884: Show separator between 440 and 490 fields
This adds a ; as separator between 440 and 490 fields used in the
same record. Currently we loop first 440 and then 490. 490 is displayed
depending on the indicator.

So after we finish the first loop, we check for existence of 490 on the
same record now to decide if a separator is needed.

To test:
- Look at various records using one or multiple 440 and 490 fields
- Verify the display is correct, especially when both appar on the
  same record in both OPAC and staff

Signed-off-by: Hayley Mapley <hayleymapley@catalyst.net.nz>
Signed-off-by: Hayley Mapley <hayleymapley@catalyst.net.nz>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>