Bug 23307: (RM follow-up) Fix breaking change to columns_settings.inc
The changes to columns_settings.inc in bug 23307 inadvertantly change
the return from an updated jQuery datatable object to a plain jQuery
object and as such broke a number of tables in Koha.  This patch simply
reverts the change whilst continuing to allow the new functionality to
work as expeced.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>