Bug 23448: Clean up subscription detail template
This patch makes a number of corrections to the subscription detail
page: Re-indented consistently with four spaces. Removed invalid
attributes. Added missing </div>.

To test, apply the patch and go to Serials. Search for a subscription
and view the details page.

The page should look correct. Validating the source should return no
errors.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>