Bug 14862: Upgrade jQuery from 1.7 to 3.4.1 in OPAC
This patch upgrades jQuery to 3.4.1, jQuery-UI to 1.12.1, and adds the
jQuery-migrate plugin to ensure backwards-compatibility with existing
jQuery plugins and code. An updated of jquery-ui-rtl.css has been
created by converting the new version of jquery-ui.css.

All jQuery assets are now include the version number in the file name
just as we now do in the staff client.

Besides updating file names in the templates, there was only one change
made: opac-results.tt had a typo which has been corrected.

To test, apply the patch and test as many different pages in the OPAC as
possible, including self-checkout and self checkin. Keep the browser
console open and watch for JavaScript errors. All JavaScript-driven
behavior should work correctly. For instance:

 - Tabs
 - Datepickers
 - Select all/none operations
 - Cart and lists popups
 - Search result highlighting

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>