Bug 23512: Reindent notices and slips page (letter.tt)
This patch re-indents the template for notices and slips. It
makes only whitespace changes. It should have no effect on the behavior
of the page.

To test, go to Tools -> Notices & slips.

Test every aspect of the notice creation and edit process. At each step
the page should work correctly.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>