Bug 23611: (follow-up) Correct ID of table
The DataTable configuration wasn't working on the table showing an
individual category's authorized values because of a mismatch in the
table id. This patch corrects it.

To test, apply the patch and view the "Authorized values" page.

The initial list of categories should be displayed as a DataTable. Click
through to see the values defined for any category. The resulting table
of values should also be a functioning DataTable.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>