Bug 23778: Regression: Guarantor entry section no longer has a unique id
This patch adds back a fieldset ID which was accidentally removed by Bug
14570:

<fieldset id="memberentry_guarantor" class="rows">

This ID is important if libraries want to customize the patron entry
page to hide the guarantor section.

To test, apply the patch and go to Patrons -> New patron.

Inspect the markup and look at the fieldset labeled "Guarantor
information." It should have the correct ID attribute.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>