Bug 23528: Show 'log in to add tags' link on all search result entries
This patch modifies the OPAC search result template so that each result
shows the "Log in to add tags" link instead of just the first.

To test, apply the patch and make sure TagsEnabled and TagsInputOnList
are both enabled.

Perform a search in the OPAC while not logged in. In the list of
results, the "Log in to add tags" link should appear with every result.

Log in to the OPAC and confirm that the "Add tag" link appears as
expected.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>