Bug 24169: Advanced editor: icons/buttons for sorting the search results are missing
This patch "manually" adds the "dataTable" class to the table of search
results in the advanced cataloging editor. This allows the correct
styles to get picked up from existing CSS.

To test, go to Cataloging and open the advanced editor. Perform a search
using the sidebar search form. In the table of results, confirm that
table sorting indicators look correct and that sorting works correctly.

Signed-off-by: Cori Lynn Arnold <carnold@dgiinc.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Joy Nelson <joy@bywatersolutions.com>