Bug 24316: Remove obsolete sample data for MANUAL_INV authorised value
Bug 23049 replaced the MANUAL_INV authorised value by a dedicated table,
but missed to remove the authorised value sample data from the non-English
installers.

This patch corrects this oversight.

Note:
The German (de-DE) file was changed a bit more, adding 2 missing translations
for the new RETURN_CLAIM_RESOLUTION and to match the sequence of the en
file.

To test:
- Verify the SQL file changes are valid by loading those files OR
- Run the installers for each language on an empty database
- Verify there is no error loading the optional authorised value data

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>

Tested with 24317 and a small followup for this bug for es-ES

Tested doing clean install for de-DE es-ES fr-CA it-IT nb-NO pl-PL
up to loading of all mandatory & optional data.
All languages load without problems.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Joy Nelson <joy@bywatersolutions.com>